+++ src/helper/api/IRecommendWords.ts
@@ -1,8 +1,3 @@
-/**
- *
- * API 서버에서 다루는 타입들을 정의 : 성격이 같은 타입들 위주로 정의할 것
- */
-
 export interface IRecommendWords {
   id: number;
   goalTitle: string;
