+++ src/helper/api/recommendWordsAPI.ts
@@ -1,6 +1,6 @@
 // import axios from 'axios';
 import client from '../../helper/clients';
-import {IResponseType} from '../../types/IRecommendWords';
+import {IResponseType} from './IRecommendWords';
 
 export default async function getRecommendWords() {
   const response = await client.get<IResponseType>('/api/v1/goals');
